--- a/frappe/__init__.py
+++ b/frappe/__init__.py
@@ -875,6 +875,7 @@ def clear_cache(user: str | None = None, doctype: str | None = None):
 	:param doctype: If doctype is given, only DocType cache is cleared."""
 	import frappe.cache_manager
 	import frappe.utils.caching
+	from frappe.website.page_renderers.document_page import clear_routing_cache
 
 	if doctype:
 		frappe.cache_manager.clear_doctype_cache(doctype)
@@ -902,6 +903,8 @@ def clear_cache(user: str | None = None, doctype: str | None = None):
 		del local.system_settings
 	if hasattr(local, "website_settings"):
 		del local.website_settings
+
+	clear_routing_cache()
 
 
 def only_has_select_perm(doctype, user=None, ignore_permissions=False):