--- a/frappe/website/page_renderers/document_page.py
+++ b/frappe/website/page_renderers/document_page.py
@@ -1,5 +1,6 @@
 import frappe
 from frappe.model.document import get_controller
+from frappe.utils.caching import redis_cache
 from frappe.website.page_renderers.base_template_page import BaseTemplatePage
 from frappe.website.router import (
 	get_doctypes_with_web_view,
@@ -22,22 +23,9 @@ class DocumentPage(BaseTemplatePage):
 		return False
 
 	def search_in_doctypes_with_web_view(self):
-		for doctype in get_doctypes_with_web_view():
-			filters = dict(route=self.path)
-			meta = frappe.get_meta(doctype)
-			condition_field = self.get_condition_field(meta)
-
-			if condition_field:
-				filters[condition_field] = 1
-
-			try:
-				self.docname = frappe.db.get_value(doctype, filters, "name")
-				if self.docname:
-					self.doctype = doctype
-					return True
-			except Exception as e:
-				if not frappe.db.is_missing_column(e):
-					raise e
+		if document := _find_matching_document_webview(self.path):
+			self.doctype, self.docname = document
+			return True
 
 	def search_web_page_dynamic_routes(self):
 		d = get_page_info_from_web_page_with_dynamic_routes(self.path)
@@ -83,7 +71,8 @@ class DocumentPage(BaseTemplatePage):
 			if prop not in self.context:
 				self.context[prop] = getattr(self.doc, prop, False)
 
-	def get_condition_field(self, meta):
+	@staticmethod
+	def get_condition_field(meta):
 		condition_field = None
 		if meta.is_published_field:
 			condition_field = meta.is_published_field
@@ -92,3 +81,26 @@ class DocumentPage(BaseTemplatePage):
 			condition_field = controller.website.condition_field
 
 		return condition_field
+
+
+@redis_cache(ttl=60 * 60)
+def _find_matching_document_webview(route: str) -> tuple[str, str] | None:
+	for doctype in get_doctypes_with_web_view():
+		filters = dict(route=route)
+		meta = frappe.get_meta(doctype)
+		condition_field = DocumentPage.get_condition_field(meta)
+
+		if condition_field:
+			filters[condition_field] = 1
+
+		try:
+			docname = frappe.db.get_value(doctype, filters, "name")
+			if docname:
+				return (doctype, docname)
+		except Exception as e:
+			if not frappe.db.is_missing_column(e):
+				raise e
+
+
+def clear_routing_cache():
+	_find_matching_document_webview.clear_cache()