--- a/frappe/website/page_renderers/document_page.py
+++ b/frappe/website/page_renderers/document_page.py
@@ -103,4 +103,7 @@ def _find_matching_document_webview(route: str) -> tuple[str, str] | None:
 
 
 def clear_routing_cache():
+	from frappe.website.doctype.web_page.web_page import get_dynamic_web_pages
+
 	_find_matching_document_webview.clear_cache()
+	get_dynamic_web_pages.clear_cache()