--- a/frappe/website/utils.py
+++ b/frappe/website/utils.py
@@ -360,7 +360,7 @@ def get_html_content_based_on_type(doc, fieldname, content_type):
 def clear_cache(path=None):
 	"""Clear website caches
 	:param path: (optional) for the given path"""
-	from frappe.website.page_renderers.document_page import clear_routing_cache
+	from frappe.website.router import clear_routing_cache
 
 	for key in ("website_generator_routes", "website_pages", "website_full_index", "sitemap_routes"):
 		frappe.cache().delete_value(key)