Context / closes: https://github.com/frappe/frappe/issues/18047


Before:
```
λ wrk -t10 -c20 -d30s http://erp.localhost:8000/
Running 30s test @ http://erp.localhost:8000/
  10 threads and 20 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   516.99ms   70.62ms 750.57ms   72.31%
    Req/Sec     4.51      2.94    10.00     73.74%
  1152 requests in 30.03s, 160.90MB read
Requests/sec:     38.36
Transfer/sec:      5.36MB
```

After:
```
λ wrk -t10 -c20 -d30s http://erp.localhost:8000/
Running 30s test @ http://erp.localhost:8000/
  10 threads and 20 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   428.96ms   60.97ms 680.33ms   69.19%
    Req/Sec     5.35      2.91    10.00     64.72%
  1389 requests in 30.03s, 194.00MB read
Requests/sec:     46.26
Transfer/sec:      6.46MB
```


Primarily achieved by caching following queries.

- Web view evaluations (we evaluate EACH webview doctype)
- Dynamic web page list, published web form list.

Note: This doesn't improve `/app` as it's already hardcoded to avoid this routing dance. https://github.com/frappe/frappe/pull/17891